pyproject.toml
[project]
name = "shoe-advisor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.9.0.post1",
    "numpy>=2.2.1",
    "openai>=1.59.4",
    "python-dotenv>=1.0.1",
]
